feat: moving more of the functions out of pfe-accordion.ts
